requirements.txt
Django>=4.2
djangorestframework
psycopg[binary]
gunicorn
jinja2
